refactor(location) : devide files by functions
