+++ src/main/java/baseball/Player.java
@@ -1,47 +1,13 @@
 package baseball;
 
 import java.util.*;
-
+import static baseball.Function.*;
 import camp.nextstep.edu.missionutils.Console;
 
 public class Player {
     protected static List<Integer> Answer;
 
     public Player() {}
-
-    public static boolean isInteger(String s){
-        try  {
-            Integer.parseInt(s);
-            return true;
-        } catch (NumberFormatException e){
-            return false;
-        }
-    }
-
-    public static boolean isPositiveNumber(String s){
-        boolean result = true;
-        if (Integer.parseInt(s) <= 0){
-            result = false;
-        }
-        return result;
-    }
-
-    public static List<Integer> changeStringtoList(String s){
-        List<Integer> result = new ArrayList<>();
-        for (int i=0; i<s.length(); i++){
-            result.add(Character.getNumericValue(s.charAt(i)));
-        }
-        return result;
-    }
-
-    public static boolean isDifferentNumbers(List<Integer> list){
-        boolean result = true;
-        Set<Integer> set = new HashSet<>(list);
-        if(set.size() != list.size()){
-            result = false;
-        }
-        return result;
-    }
 
     public static void isValidAnswer(String input){
         if(input.length() != 3 || !isInteger(input) || !isPositiveNumber(input) || !isDifferentNumbers(changeStringtoList(input))){
@@ -57,10 +23,5 @@ public class Player {
         } catch (IllegalArgumentException e) {
             System.out.println(e.getMessage());
         }
-
     }
-
-
-
-
 }
