+++ src/test/java/baseball/ComputerTest.java
@@ -1,0 +1,16 @@
+package baseball;
+
+import org.junit.jupiter.api.Test;
+import java.util.*;
+import static baseball.Computer.*;
+import static org.assertj.core.api.Assertions.assertThat;
+
+public class ComputerTest {
+
+    @Test
+    void makeAnswer_컴퓨터_정답_만들기(){
+        List<Integer> answer = makeAnswer();
+        int result = answer.size();
+        assertThat(result).isEqualTo(3);
+    }
+}
