+++ src/test/java/baseball/FunctionTest.java
@@ -2,8 +2,7 @@ package baseball;
 
 import org.junit.jupiter.api.Test;
 import java.util.*;
-import static baseball.Computer.*;
-import static baseball.Player.*;
+import static baseball.Function.*;
 import static org.assertj.core.api.Assertions.*;
 
 public class FunctionTest {
@@ -23,13 +22,6 @@ public class FunctionTest {
             }
         }
         assertThat(result).isEqualTo(true);
-    }
-
-    @Test
-    void makeAnswer_컴퓨터_정답_만들기(){
-        List<Integer> answer = makeAnswer();
-        int result = answer.size();
-        assertThat(result).isEqualTo(3);
     }
 
     @Test
@@ -55,27 +47,5 @@ public class FunctionTest {
         List<Integer> list = List.of(1,2,4,4);
         boolean result = isDifferentNumbers(list);
         assertThat(result).isEqualTo(false);
-    }
-
-    @Test
-    void isValidAnswer_메서드_사용시_입력받은_문자열이_조건을_만족하지_못할시_예외_발생() {
-        String input = "1234";
-
-        assertThatThrownBy(() -> isValidAnswer(input))
-                .isInstanceOf(IllegalArgumentException.class)
-                .hasMessageContaining("게임종료");
-    }
-
-    @Test
-    void guessAnswer_정답_예측_메소드(){
-        Player player = new Player();
-        String input = "123";
-        try {
-            isValidAnswer(input);
-            player.Answer = changeStringtoList(input);
-        } catch (IllegalArgumentException e) {
-            System.out.println(e.getMessage());
-        }
-        assertThat(player.Answer).isEqualTo(List.of(1,2,3));
     }
 }
